requirements.txt
# installs dependencies from ./data-tools/setup.py, and the package itself,
-e ./data-tools/
-e ./fhg-utils/

jupyter
ipython
matplotlib
seaborn
tabulate
mlflow[extras]

-i https://pypi.python.org/simple
absl-py==0.10.0
astunparse==1.6.3
cachetools==4.1.1; python_version ~= '3.5'
certifi==2020.6.20
chardet==3.0.4
gast==0.3.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth-oauthlib==0.4.1
google-auth==1.21.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==1.7.0; python_version < '3.8'
joblib==0.16.0; python_version >= '3.6'
keras-preprocessing==1.1.2
markdown==3.2.2; python_version >= '3.5'
numpy==1.18.5
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
opt-einsum==3.3.0; python_version >= '3.5'
pandas==1.1.1
protobuf==3.13.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2020.1
requests-oauthlib==1.3.0
requests==2.24.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rsa==4.6; python_version >= '3.5'
scikit-learn==0.23.2; python_version >= '3.6'
scipy==1.4.1; python_version >= '3.5'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sklearn==0.0
tensorboard-plugin-wit==1.7.0
tensorboard==2.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1'
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0; python_version >= '3.5'
urllib3==1.25.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wheel==0.35.1; python_version >= '3'
wrapt==1.12.1
zipp==3.1.0; python_version >= '3.6'
